+++ example/lib/enemy/goblin.dart
@@ -11,25 +11,25 @@ class Goblin extends SimpleEnemy {
 
   Goblin(Position initPosition)
       : super(
-            animationIdleRight: FlameAnimation.Animation.sequenced(
+            animIdleRight: FlameAnimation.Animation.sequenced(
               "enemy/goblin_idle.png",
               6,
               textureWidth: 16,
               textureHeight: 16,
             ),
-            animationIdleLeft: FlameAnimation.Animation.sequenced(
+            animIdleLeft: FlameAnimation.Animation.sequenced(
               "enemy/goblin_idle_left.png",
               6,
               textureWidth: 16,
               textureHeight: 16,
             ),
-            animationRunRight: FlameAnimation.Animation.sequenced(
+            animRunRight: FlameAnimation.Animation.sequenced(
               "enemy/goblin_run_right.png",
               6,
               textureWidth: 16,
               textureHeight: 16,
             ),
-            animationRunLeft: FlameAnimation.Animation.sequenced(
+            animRunLeft: FlameAnimation.Animation.sequenced(
               "enemy/goblin_run_left.png",
               6,
               textureWidth: 16,